Clean up scraper and modify logic to have assumption that there is no need to rescrape stuff midway through..
I need to think of a robust way to update my database whenever we are scraping new courses into the database.

The main thing I have been considering is what happens when a scrape fails midway through when working with a production database — I will need a way to rescrape the other ones (but this is actually easy to fix since I can just check for the PKs and retry again — don’t even need snapshots).

I think the sensible thing to do for the prod database is to just make sure the scrape is successful locally before trying it on the actual db. If something bad does happen, then we can just purge those courseOfferings by targetting those with that termid  — we can have a snapshot before we make any changes as well just in case.