Add professor linking
This commit is a working version for the actuarial science department. It should also work for all other departments but this is expensive to run, so I haven't checked. The main goal from here is to replace the use of the join tables so that we can simplify our business logic in the future and then to optimize the llm as much as possible to save on tokens and cost where possible